requirements.txt
streamlit
transformers
tensorflow
google-generativeai
astor
numpy
pandas
scikit-learn
torch
datasets
huggingface-hub
